refactor: rename generate_HRD_associated_func to satisfy snake case notation
