+++ src/generate_hrd_associated_funcs.rs
